Update README.md
# Unsupervised-Learning
* Domain - Automotive
* focus - Incentivize drivers 

# Business Challenge/requirement

Company Prajwalxyz is the largest provider of electric vehicle (e-vehicle) betteries. It provides battery on  a rental model to e-vehicle drivers.
Drivers rent battery typically for a day and then replace it with a charged battery from the company.

Company Prajwalxyz has a variable pricing model based on driver's driving history. As the life of a battery depends on factors such as over speeding,
distance driven per day etc.

You as a ML expert have to create a cluster model where drivers cab be grouped together based on the driving data.